Start addressing reviewer comments
