+++ src/test/java/org/broadinstitute/hellbender/tools/sv/SVTestUtils.java
@@ -3,54 +3,54 @@ package org.broadinstitute.hellbender.tools.sv;
 import com.google.common.collect.Lists;
 import htsjdk.samtools.SAMSequenceDictionary;
 import htsjdk.samtools.SAMSequenceRecord;
+import htsjdk.samtools.reference.ReferenceSequenceFile;
 import htsjdk.variant.variantcontext.*;
+import org.broadinstitute.hellbender.GATKBaseTest;
+import org.broadinstitute.hellbender.engine.GATKPath;
 import org.broadinstitute.hellbender.testutils.VariantContextTestUtils;
 import org.broadinstitute.hellbender.tools.spark.sv.utils.GATKSVVCFConstants;
-import org.broadinstitute.hellbender.tools.sv.cluster.LocatableClusterEngine;
-import org.broadinstitute.hellbender.tools.sv.cluster.SVClusterEngine;
-import org.broadinstitute.hellbender.tools.sv.cluster.SVCollapser;
+import org.broadinstitute.hellbender.tools.sv.cluster.*;
 import org.broadinstitute.hellbender.utils.GenomeLoc;
 import org.broadinstitute.hellbender.utils.GenomeLocParser;
+import org.broadinstitute.hellbender.utils.reference.ReferenceUtils;
 import org.testng.Assert;
 import org.testng.TestException;
 
 import java.util.*;
-import java.util.function.Function;
+import java.util.stream.Collectors;
 
 public class SVTestUtils {
     public final static int chr1Length = 249250621;
 
+    public static final ReferenceSequenceFile ref = ReferenceUtils.createReferenceReader(new GATKPath(GATKBaseTest.hg38Reference));
     public final static SAMSequenceDictionary dict = new SAMSequenceDictionary(
             Arrays.asList(new SAMSequenceRecord("chr1", chr1Length),
                     new SAMSequenceRecord("chr10", 135534747),
                     new SAMSequenceRecord("chrX", 155270560)));
 
     private final static GenomeLocParser glParser = new GenomeLocParser(SVTestUtils.dict);
-    public static final Function<Collection<SVCallRecord>, SVCallRecord> defaultCollapser =
-            new SVCollapser(SVCollapser.BreakpointSummaryStrategy.MEDIAN_START_MEDIAN_END)::collapse;
+    public static final SVCollapser<SVCallRecord> defaultCollapser =
+            new CanonicalSVCollapser(ref, CanonicalSVCollapser.BreakpointSummaryStrategy.MEDIAN_START_MEDIAN_END);
+
+    public static CanonicalSVLinkage<SVCallRecord> getNewDefaultLinkage() {
+        final CanonicalSVLinkage<SVCallRecord> linkage = new CanonicalSVLinkage<>(SVTestUtils.dict, false);
+        linkage.setDepthOnlyParams(defaultDepthOnlyParameters);
+        linkage.setMixedParams(defaultMixedParameters);
+        linkage.setEvidenceParams(defaultEvidenceParameters);
+        return linkage;
+    }
 
     public static SVClusterEngine<SVCallRecord> getNewDefaultSingleLinkageEngine() {
-        final SVClusterEngine<SVCallRecord> engine = new SVClusterEngine<>(SVTestUtils.dict, LocatableClusterEngine.CLUSTERING_TYPE.SINGLE_LINKAGE, false, defaultCollapser);
-        engine.setDepthOnlyParams(defaultDepthOnlyParameters);
-        engine.setMixedParams(defaultMixedParameters);
-        engine.setEvidenceParams(defaultEvidenceParameters);
-        return engine;
+        return new SVClusterEngine<>(SVClusterEngine.CLUSTERING_TYPE.SINGLE_LINKAGE, defaultCollapser, getNewDefaultLinkage());
     }
 
     public static SVClusterEngine<SVCallRecord> getNewDefaultMaxCliqueEngine() {
-        final SVClusterEngine<SVCallRecord> engine = new SVClusterEngine<>(SVTestUtils.dict, LocatableClusterEngine.CLUSTERING_TYPE.MAX_CLIQUE, false, defaultCollapser);
-        engine.setDepthOnlyParams(defaultDepthOnlyParameters);
-        engine.setMixedParams(defaultMixedParameters);
-        engine.setEvidenceParams(defaultEvidenceParameters);
-        return engine;
+        return new SVClusterEngine<>(SVClusterEngine.CLUSTERING_TYPE.MAX_CLIQUE, defaultCollapser, getNewDefaultLinkage());
     }
 
-    protected static final SVClusterEngine.ClusteringParameters defaultDepthOnlyParameters =
-            new SVClusterEngine.DepthClusteringParameters(0.8, 0, 0);
-    protected static final SVClusterEngine.ClusteringParameters defaultMixedParameters =
-            new SVClusterEngine.MixedClusteringParameters(0.8, 1000, 0);
-    protected static final SVClusterEngine.ClusteringParameters defaultEvidenceParameters =
-            new SVClusterEngine.EvidenceClusteringParameters(0.5, 500, 0);
+    public static final ClusteringParameters defaultDepthOnlyParameters = ClusteringParameters.createDepthParameters(0.8, 0, 0);
+    public static final ClusteringParameters defaultMixedParameters = ClusteringParameters.createMixedParameters(0.8, 1000, 0);
+    public static final ClusteringParameters defaultEvidenceParameters = ClusteringParameters.createPesrParameters(0.5, 500, 0);
 
     public static final SVClusterEngine<SVCallRecord> defaultSingleLinkageEngine = getNewDefaultSingleLinkageEngine();
     public static final SVClusterEngine<SVCallRecord> defaultMaxCliqueEngine = getNewDefaultMaxCliqueEngine();
@@ -59,14 +59,14 @@ public class SVTestUtils {
 
     public final static int length = 10000;
 
-    public final static int length1b = (int)Math.round(defaultSingleLinkageEngine.getDepthOnlyParams().getReciprocalOverlap() * length);
+    public final static int length1b = (int)Math.round(defaultDepthOnlyParameters.getReciprocalOverlap() * length);
 
     public final static int start1b = start + length - length1b;
 
     //separated from end of call1 by defragmenter padding (in bin space, according to bins defined by targetIntervals below)
     public final static int start2 = start + length*14/9;
 
-    public final static int start3 = start + (int)Math.round(Math.floor((1- defaultSingleLinkageEngine.getDepthOnlyParams().getReciprocalOverlap())*length));
+    public final static int start3 = start + (int)Math.round(Math.floor((1- defaultDepthOnlyParameters.getReciprocalOverlap())*length));
 
     //make intervals like xxxx----xxxx----xxxx----xxxx----xxxx
     public final static List<GenomeLoc> targetIntervals = new ArrayList<>(
@@ -92,121 +92,155 @@ public class SVTestUtils {
                     glParser.createGenomeLoc("chr10", start, start + length - 1)))
             ;
 
-    public final static Genotype sample1 = new GenotypeBuilder("sample1",
+    public final static GenotypeBuilder sample1 = new GenotypeBuilder("sample1",
             Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL))
-            .attribute(GATKSVVCFConstants.COPY_NUMBER_FORMAT, 1).make();
+            .attribute(GATKSVVCFConstants.COPY_NUMBER_FORMAT, 1);
 
-    public final static Genotype sample1Ploidy1 = new GenotypeBuilder("sample1",
+    public final static GenotypeBuilder sample1Ploidy1 = new GenotypeBuilder("sample1",
             Lists.newArrayList(Allele.SV_SIMPLE_DEL))
-            .attribute(GATKSVVCFConstants.COPY_NUMBER_FORMAT, 1).make();
+            .attribute(GATKSVVCFConstants.COPY_NUMBER_FORMAT, 1);
 
-    public final static Genotype sample1_CN0 = new GenotypeBuilder("sample1",
+    public final static GenotypeBuilder sample1_CN0 = new GenotypeBuilder("sample1",
             Lists.newArrayList(Allele.SV_SIMPLE_DEL, Allele.SV_SIMPLE_DEL))
-            .attribute(GATKSVVCFConstants.COPY_NUMBER_FORMAT, 0).make();
+            .attribute(GATKSVVCFConstants.COPY_NUMBER_FORMAT, 0);
 
-    public final static Genotype sample2 = GenotypeBuilder.create("sample2",
-            Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DUP),
-            Collections.singletonMap(GATKSVVCFConstants.COPY_NUMBER_FORMAT, 3));
+    public final static GenotypeBuilder sample2 = new GenotypeBuilder("sample2",
+            Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DUP))
+            .attribute(GATKSVVCFConstants.COPY_NUMBER_FORMAT, 3);
 
-    public final static Genotype sample3 = GenotypeBuilder.create("sample3",
-            Lists.newArrayList(Allele.SV_SIMPLE_DEL, Allele.SV_SIMPLE_DEL),
-            Collections.singletonMap(GATKSVVCFConstants.COPY_NUMBER_FORMAT, 0));
+    public final static GenotypeBuilder sample3 = new GenotypeBuilder("sample3",
+            Lists.newArrayList(Allele.SV_SIMPLE_DEL, Allele.SV_SIMPLE_DEL))
+            .attribute(GATKSVVCFConstants.COPY_NUMBER_FORMAT, 0);
 
-    public final static SVCallRecord rightEdgeCall = new SVCallRecord("rightEdgeCall", "chr1", chr1Length - 99, true,
+    public final static SVCallRecord makeRecord(final String id,
+                                                 final String contigA,
+                                                 final int positionA,
+                                                 final boolean strandA,
+                                                 final String contigB,
+                                                 final int positionB,
+                                                 final boolean strandB,
+                                                 final StructuralVariantType type,
+                                                 final int length,
+                                                 final List<String> algorithms,
+                                                 final List<Allele> alleles,
+                                                 final List<GenotypeBuilder> genotypeBuilders) {
+        final Allele refAllele = Allele.create(ReferenceUtils.getRefBaseAtPosition(ref, contigA, positionA));
+        final List<Allele> newAlleles = replaceRefNAlleles(alleles, refAllele);
+        for (final GenotypeBuilder builder : genotypeBuilders) {
+            builder.alleles(replaceRefNAlleles(builder.make().getAlleles(), refAllele));
+        }
+        final List<Genotype> genotypes = genotypeBuilders.stream().map(GenotypeBuilder::make).collect(Collectors.toList());
+        return new SVCallRecord(id, contigA, positionA, strandA, contigB, positionB, strandB, type, length, algorithms,
+                newAlleles, genotypes);
+    }
+
+    private static final List<Allele> replaceRefNAlleles(final List<Allele> alleles, final Allele newRefAllele) {
+        return alleles.stream()
+                .map(a -> a.equals(Allele.REF_N) ? newRefAllele : a)
+                .collect(Collectors.toList());
+    }
+
+    public static final Genotype makeGenotypeWithRefAllele(final GenotypeBuilder builder, final String contig, final int position) {
+        final Allele refAllele = Allele.create(ReferenceUtils.getRefBaseAtPosition(ref, contig, position));
+        builder.alleles(replaceRefNAlleles(builder.make().getAlleles(), refAllele));
+        return builder.make();
+    }
+
+    public final static SVCallRecord rightEdgeCall = makeRecord("rightEdgeCall", "chr1", chr1Length - 99, true,
             "chr1", chr1Length, true,
             StructuralVariantType.CNV, 100,
             Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
             Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL, Allele.SV_SIMPLE_DUP),
             Arrays.asList(sample1, sample2));
 
-    public final static SVCallRecord leftEdgeCall = new SVCallRecord("leftEdgeCall", "chr1", 1, true,
+    public final static SVCallRecord leftEdgeCall = makeRecord("leftEdgeCall", "chr1", 1, true,
             "chr1", length/2, true,
             StructuralVariantType.CNV, length/2,
             Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
             Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL),
             Arrays.asList(sample1, sample2));
 
-    public final static SVCallRecord nonDepthOnly = new SVCallRecord("nonDepthOnly", "chr1", start, true,
+    public final static SVCallRecord nonDepthOnly = makeRecord("nonDepthOnly", "chr1", start, true,
             "chr1", start + length -1, true,
             StructuralVariantType.CNV, length,
             Arrays.asList(GATKSVVCFConstants.DEPTH_ALGORITHM, "PE"),
             Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL, Allele.SV_SIMPLE_DUP),
             Arrays.asList(sample1, sample2));
 
-    public final static SVCallRecord call1 = new SVCallRecord("call1", "chr1", start, true,
+    public final static SVCallRecord call1 = makeRecord("call1", "chr1", start, true,
             "chr1", start + length -1, true,
             StructuralVariantType.CNV, length,
             Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
             Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL, Allele.SV_SIMPLE_DUP),
             Arrays.asList(sample1, sample2));
 
-    public final static SVCallRecord call1b = new SVCallRecord("call1b", "chr1", start1b, true,
+    public final static SVCallRecord call1b = makeRecord("call1b", "chr1", start1b, true,
             "chr1", start1b + length1b + 1, true,
             StructuralVariantType.CNV, length1b,
             Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
             Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL, Allele.SV_SIMPLE_DUP),
             Arrays.asList(sample1, sample2));
 
-    public final static SVCallRecord call2 = new SVCallRecord("call2", "chr1", start2, true,
+    public final static SVCallRecord call2 = makeRecord("call2", "chr1", start2, true,
             "chr1", start2 + length -1, true,
             StructuralVariantType.CNV, length,
             Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
             Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL, Allele.SV_SIMPLE_DUP),
             Arrays.asList(sample1, sample2));
 
-    public final static SVCallRecord call3 = new SVCallRecord("call3", "chr1", start2+length, true,
+    public final static SVCallRecord call3 = makeRecord("call3", "chr1", start2+length, true,
             "chr1", start2 + length + length -1, true,
             StructuralVariantType.CNV, length,
             Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
             Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL, Allele.SV_SIMPLE_DUP),
             Arrays.asList(sample1, sample2));
 
-    public final static SVCallRecord call4_chr10 = new SVCallRecord("call4_chr10", "chr10", start, true,
+    public final static SVCallRecord call4_chr10 = makeRecord("call4_chr10", "chr10", start, true,
             "chr10", start + length -1, true,
             StructuralVariantType.CNV, length,
             Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
             Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL, Allele.SV_SIMPLE_DUP),
             Arrays.asList(sample1, sample2));
 
-    public final static SVCallRecord call1_CN1 = new SVCallRecord("call1_CN1", "chr1", start, true,
+    public final static SVCallRecord call1_CN1 = makeRecord("call1_CN1", "chr1", start, true,
             "chr1", start + length -1, true,
             StructuralVariantType.CNV, length,
             Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
             Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL),
             Arrays.asList(sample1));
 
-    public final static SVCallRecord call2_CN0 = new SVCallRecord("call2_CN0", "chr1", start2, true,
+    public final static SVCallRecord call2_CN0 = makeRecord("call2_CN0", "chr1", start2, true,
             "chr1", start2 + length -1, true,
             StructuralVariantType.CNV, length,
             Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
             Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL),
             Arrays.asList(sample1_CN0));
 
-    public final static SVCallRecord sameBoundsSampleMismatch = new SVCallRecord("sameBoundsSampleMismatch", "chr1", start, true,
+    public final static SVCallRecord sameBoundsSampleMismatch = makeRecord("sameBoundsSampleMismatch", "chr1", start, true,
             "chr1", start + length -1, true,
             StructuralVariantType.CNV, length,
             Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
             Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL),
             Arrays.asList(sample1, sample3));
 
-    public final static List<Genotype> threeGenotypes = Arrays.asList(sample1, sample2, sample3);
+    public final static List<Genotype> threeGenotypes = Arrays.asList(sample1.make(), sample2.make(), sample3.make());
 
-    public final static SVCallRecord inversion = new SVCallRecord("inversion", "chr1", 10000, true, "chr1", 20000, true,
+    public final static SVCallRecord inversion = makeRecord("inversion", "chr1", 10000, true, "chr1", 20000, true,
             StructuralVariantType.INV, 10001, Arrays.asList("SR", "PE"), Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DUP), Collections.singletonList(sample2));
 
-    public static final SVCallRecord depthOnly = new SVCallRecord("depthOnly", "chr1", 10000, true, "chr1", 20000, true,
+    public static final SVCallRecord depthOnly = makeRecord("depthOnly", "chr1", 10000, true, "chr1", 20000, true,
             StructuralVariantType.CNV, 10001, Collections.singletonList("depth"), Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL), Collections.singletonList(sample1));
 
-    public static final SVCallRecord depthAndStuff = new SVCallRecord("depthAndStuff", "chr1", 10000, true, "chr1", 20000, true,
+    public static final SVCallRecord depthAndStuff = makeRecord("depthAndStuff", "chr1", 10000, true, "chr1", 20000, true,
                     StructuralVariantType.CNV, 10001, Arrays.asList("depth", "PE"), Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DUP), Collections.singletonList(sample2));
 
-    public static final SVCallRecord depthAndStuff2 = new SVCallRecord("depthAndStuff2", "chr1", 10000 - defaultSingleLinkageEngine.getEvidenceParams().getWindow(), true, "chr1", 20000, true,
-            StructuralVariantType.CNV, 10001 + defaultSingleLinkageEngine.getEvidenceParams().getWindow(), Arrays.asList("depth", "PE"), Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DUP), Collections.singletonList(sample2));
+    public static final SVCallRecord depthAndStuff2 = makeRecord("depthAndStuff2", "chr1", 10000 - defaultEvidenceParameters.getWindow(), true, "chr1", 20000, true,
+            StructuralVariantType.CNV, 10001 + defaultEvidenceParameters.getWindow(), Arrays.asList("depth", "PE"), Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DUP), Collections.singletonList(sample2));
 
-    public static final SVCallRecord depthAndStuff3 = new SVCallRecord("depthAndStuff3", "chr1", 10000 + defaultSingleLinkageEngine.getEvidenceParams().getWindow(), true, "chr1", 20000, true,
-            StructuralVariantType.CNV, 10001 - defaultSingleLinkageEngine.getEvidenceParams().getWindow(), Arrays.asList("depth", "PE"), Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DUP), Collections.singletonList(sample2));
+    public static final SVCallRecord depthAndStuff3 = makeRecord("depthAndStuff3", "chr1", 10000 + defaultEvidenceParameters.getWindow(), true, "chr1", 20000, true,
+            StructuralVariantType.CNV, 10001 - defaultEvidenceParameters.getWindow(), Arrays.asList("depth", "PE"), Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DUP), Collections.singletonList(sample2));
 
-    public final static SVCallRecord overlapsCall1 = new SVCallRecord("overlapsCall1", "chr1", start3, true,
+    public final static SVCallRecord overlapsCall1 = makeRecord("overlapsCall1", "chr1", start3, true,
             "chr1", start3 + length -1, true,
             StructuralVariantType.CNV, length,
             Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
