+++ src/test/java/org/broadinstitute/hellbender/tools/sv/cluster/CNVDefragmenterTest.java
@@ -17,7 +17,7 @@ import java.util.Collections;
 public class CNVDefragmenterTest {
 
     private final SAMSequenceDictionary dictionary = SVTestUtils.dict;
-    private final CNVDefragmenter defragmenter = new CNVDefragmenter(dictionary, 0.5, 0.6);
+    private final CNVLinkage defragmenter = new CNVLinkage(dictionary, 0.5, 0.6);
 
     @Test
     public void testClusterTogether() {
@@ -29,43 +29,43 @@ public class CNVDefragmenterTest {
                 1000, Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
                 Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DUP),
                 Collections.emptyList(), Collections.emptyMap());
-        Assert.assertFalse(defragmenter.clusterTogether(deletion, duplication), "Different sv types should not cluster");
+        Assert.assertFalse(defragmenter.areClusterable(deletion, duplication), "Different sv types should not cluster");
 
         final SVCallRecord duplicationNonDepthOnly = new SVCallRecord("test_dup", "chr1", 1000, false, "chr1", 1999, false, StructuralVariantType.DUP,
                 1000, Lists.newArrayList(GATKSVVCFConstants.DEPTH_ALGORITHM, "pesr"),
                 Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DUP),
                 Collections.emptyList(), Collections.emptyMap());
-        Assert.assertFalse(defragmenter.clusterTogether(duplication, duplicationNonDepthOnly), "Clustered records must be depth-only");
+        Assert.assertFalse(defragmenter.areClusterable(duplication, duplicationNonDepthOnly), "Clustered records must be depth-only");
 
         final SVCallRecord cnv = new SVCallRecord("test_cnv", "chr1", 1000, false, "chr1", 1999, false, StructuralVariantType.CNV,
                 1000, Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
                 Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL, Allele.SV_SIMPLE_DUP),
                 Collections.emptyList(), Collections.emptyMap());
-        Assert.assertFalse(defragmenter.clusterTogether(deletion, cnv), "Different sv types should not cluster");
+        Assert.assertFalse(defragmenter.areClusterable(deletion, cnv), "Different sv types should not cluster");
 
         final SVCallRecord insertion = new SVCallRecord("test_ins", "chr1", 1000, true, "chr1", 1001, false, StructuralVariantType.INS,
                 1000, Collections.singletonList("pesr"),
                 Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_INS),
                 Collections.emptyList(), Collections.emptyMap());
-        Assert.assertFalse(defragmenter.clusterTogether(insertion, insertion), "Only CNVs should be valid");
+        Assert.assertFalse(defragmenter.areClusterable(insertion, insertion), "Only CNVs should be valid");
 
         final SVCallRecord deletion2 = new SVCallRecord("test_del2", "chr1", 1000, true, "chr1", 1999, false, StructuralVariantType.DEL,
                 1000, Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
                 Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL),
                 Collections.emptyList(), Collections.emptyMap());
-        Assert.assertTrue(defragmenter.clusterTogether(deletion, deletion2), "Valid identical records should cluster");
+        Assert.assertTrue(defragmenter.areClusterable(deletion, deletion2), "Valid identical records should cluster");
 
         final SVCallRecord deletion3 = new SVCallRecord("test_del3", "chr1", 2999, true, "chr1", 3998, false, StructuralVariantType.DEL,
                 1000, Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
                 Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL),
                 Collections.emptyList(), Collections.emptyMap());
-        Assert.assertTrue(defragmenter.clusterTogether(deletion, deletion3), "Should cluster due to overlap");
+        Assert.assertTrue(defragmenter.areClusterable(deletion, deletion3), "Should cluster due to overlap");
 
         final SVCallRecord deletion4 = new SVCallRecord("test_del3", "chr1", 3000, true, "chr1", 3999, false, StructuralVariantType.DEL,
                 1000, Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
                 Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL),
                 Collections.emptyList(), Collections.emptyMap());
-        Assert.assertFalse(defragmenter.clusterTogether(deletion, deletion4), "Should barely not cluster");
+        Assert.assertFalse(defragmenter.areClusterable(deletion, deletion4), "Should barely not cluster");
     }
 
     @DataProvider(name = "testOverlapBuilders")
@@ -96,7 +96,7 @@ public class CNVDefragmenterTest {
     }
 
     @Test(dataProvider= "testOverlapBuilders")
-    private void testSampleOverlap(final GenotypeBuilder refGenotypeMaker,
+    public void testSampleOverlap(final GenotypeBuilder refGenotypeMaker,
                                    final GenotypeBuilder altGenotypeMaker1,
                                    final GenotypeBuilder altGenotypeMaker2,
                                    final String name) {
@@ -144,16 +144,16 @@ public class CNVDefragmenterTest {
                 Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL),
                 Lists.newArrayList(altGenotypeMaker1.name("sample1").make(), altGenotypeMaker1.name("sample2").make(), altGenotypeMaker1.name("sample3").make()),
                 Collections.emptyMap());
-        Assert.assertFalse(defragmenter.clusterTogether(deletionARR1, deletionARR2));
-        Assert.assertFalse(defragmenter.clusterTogether(deletionARR1, deletionRAR));
-        Assert.assertFalse(defragmenter.clusterTogether(deletionARR1, deletionAAR));
-        Assert.assertFalse(defragmenter.clusterTogether(deletionARR1, deletionAAA));
-        Assert.assertFalse(defragmenter.clusterTogether(deletionRAR, deletionAAR));
-        Assert.assertFalse(defragmenter.clusterTogether(deletionRAR, deletionAAA));
+        Assert.assertFalse(defragmenter.areClusterable(deletionARR1, deletionARR2));
+        Assert.assertFalse(defragmenter.areClusterable(deletionARR1, deletionRAR));
+        Assert.assertFalse(defragmenter.areClusterable(deletionARR1, deletionAAR));
+        Assert.assertFalse(defragmenter.areClusterable(deletionARR1, deletionAAA));
+        Assert.assertFalse(defragmenter.areClusterable(deletionRAR, deletionAAR));
+        Assert.assertFalse(defragmenter.areClusterable(deletionRAR, deletionAAA));
 
-        Assert.assertTrue(defragmenter.clusterTogether(deletionARR1, deletionARR1Copy));
-        Assert.assertTrue(defragmenter.clusterTogether(deletionAAR, deletionAAA));
-        Assert.assertTrue(defragmenter.clusterTogether(deletionAAR, deletionAAR2));
+        Assert.assertTrue(defragmenter.areClusterable(deletionARR1, deletionARR1Copy));
+        Assert.assertTrue(defragmenter.areClusterable(deletionAAR, deletionAAA));
+        Assert.assertTrue(defragmenter.areClusterable(deletionAAR, deletionAAR2));
     }
 
     @DataProvider(name = "maxPositionIntervals")
@@ -180,7 +180,7 @@ public class CNVDefragmenterTest {
                 call2End - call2Start + 1, Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
                 Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL),
                 Collections.emptyList(), Collections.emptyMap());
-        Assert.assertTrue(defragmenter.clusterTogether(call1, call2));
+        Assert.assertTrue(defragmenter.areClusterable(call1, call2));
 
         final int call3Start = maxClusterableStart + 1;
         final int call3End = dictionary.getSequence(call1.getContigA()).getSequenceLength();
@@ -188,7 +188,7 @@ public class CNVDefragmenterTest {
                 call3End - call3Start + 1, Collections.singletonList(GATKSVVCFConstants.DEPTH_ALGORITHM),
                 Lists.newArrayList(Allele.REF_N, Allele.SV_SIMPLE_DEL),
                 Collections.emptyList(), Collections.emptyMap());
-        Assert.assertFalse(defragmenter.clusterTogether(call1, call3));
+        Assert.assertFalse(defragmenter.areClusterable(call1, call3));
     }
 
 }
